[FIX] Changed index to disallow duplicate votes
We added `class_name: 'Profile'` to the Vote#voter relation, which causes ActiveRecord to omit the setting of the `voter_type` column. So now we have a `votes` table which contains:
`voteable_type: string, voteable_id: int, voter_id: int, voter_type: null`
Meanwhile, we create new votes with `CreateVote` where in #initialize_edge a join is done to search for an existing vote, so it's overwritten when one exists.

But since our unique constraint contains the `voter_type` column, which is now set to NULL, duplicate votes are still possible whenever they are saved multiple times because NULL never equals NULL in SQL.
This migration updates the index to only check for the (voteable_id, voteable_type, voter_id) triple, which only contains `NOT NULL` values.