[IMP] Allow registering without a password
The password will be set to a random string. The user will receive an email which is a combination of confirmation and reset_password

The user will be confirmed when updating the password, but only when he has not received a new confirmation email in the meantime. This to prevent a wrong email is to be confirmed. See http://blog.plataformatec.com.br/2013/08/devise-3-1-now-with-more-secure-defaults/