[ARGU-931] Redirect to granted object if one is present
We now redirect the user to a granted object if exactly one is present.
This enables managers to generate tokens which will redirect users to a certain forum if the token was generated specifically for that purpose.
When the invited-for group contains multiple grants, it’s impossible to deterministically redirect to the ‘correct’ object, since there is no way to
know whether higher or lower in the tree is correct.

This commit should account for a current use case which needs to be deployed quickly, but it can be removed once a more generalised system is in
place, such as TS-10 or TS-23.