[ARGU-1023] Refactor policies
Determine wether certain actions are allowed based on the {action}_roles method, which can be overridden in policies.
Remove methods that implement the same logic as the superclass.
Also rearrange alphabetically and make more methods private
No longer allow the creation of projects. This was already effectively the case because of the active_for