[ARGU-1147] Add SPI controller for FE token generation
The FE can now request tokens without using the public OAuth API.
This enables the FE to be a part of the OAuth code flow in the future (
since it handles the Argu domain authentication).

The client credentials flow has been disabled since the FE was the only
application using it. Once the classic FE has been removed, the password
flow should also be removed to increase security.

The guest token API generates its own session ids with the same
mechanism as the Rails session middleware (`SecureRandom.hex`).

Refresh token logic isn't taken into account in this implementation, but
should be handled by the FE in the future.

The changes in `resource_owner_from_credentials` keep current behaviour
for the classic FE, but return the exception object for modern FE
requests and return `nil` for external OAuth calls (which corresponds
with the expected behaviour in Doorkeeper's token_controller).