[ARGU-1231] Load styles and font asynchronously
The style link tag can't be loaded async, so we have to include it via
JavaScript. A fallback has been included in the case that no JavaScript
is available. The link `prefetch` value should direct the browser to
start fetching the css bundle before the actual style link is included
by the JS code, but not in a blocking way like with rel `preload`.

This also introduces `essential.css`, which contains basic styling for
above-the-fold rendering which is embedded into the html to increase
render speed. Since the assets pipeline isn't available in production,
the file is read in plain ruby and low-level Rails API's.