[ARGU-974] The least concern with concernables
Tries to abstract the behaviour over the models, once a
`concern <Concernable>` has been called on a model, the accompanying
collections, actions, serializers (and routes in the future) should
(near) automatically be added.

The current implementation still needs the child classes to physically
exist in addition to them needing to do some setup calls (e.g.
`include_actions`). But this should be refactored away at a later stage.